third_party/blink/web_tests/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

Tests the properties of XMLHttpRequest exceptions.

PASS e.toString() is "NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/resources/infinite-loop.php'."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) is not "[object DOMExceptionPrototype]". Was "[object DOMException]"
PASS e.constructor.toString() is "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS e.NETWORK_ERR is e.constructor.NETWORK_ERR
PASS e.NETWORK_ERR is 19
